pyproject.toml
[tool.poetry]
name = "eroerobot"
version = "0.1.0"
description = ""
authors = ["384606266 <384606266@qq.com>"]

[tool.poetry.dependencies]
python = "^3.8"
graia-ariadne = {extras = ["full"], version = "^0.5.3"}
graiax-silkcoder = {extras = ["ffmpeg"], version = "^0.1.6"}
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
httpx = "^0.22.0"
requests = "^2.27.1"
ffmpy = "^0.3.0"
graia-saya = "0.0.14"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
# 这里以清华源举例，你也可以使用其他源
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false
